+++ src/Haskellite/Types.hs
@@ -6,6 +6,8 @@
 module Haskellite.Types
   ( AppPaths (..)
   , ActivationSource (..)
+  , HotkeyKey (..)
+  , HotkeyModifiers (..)
   , HotkeyPreset (..)
   , InstallProgress (..)
   , InstallStage (..)
@@ -16,14 +18,18 @@ module Haskellite.Types
   , TranscriptSegment (..)
   , defaultSettings
   , emptyRecognitionResult
+  , hotkeyBinding
+  , validHotkeyBinding
   , renderTranscript
   ) where
 
 import Data.Aeson
   ( FromJSON (..)
   , ToJSON (..)
+  , Value (String)
   , object
   , withObject
+  , (.:)
   , (.:?)
   , (.!=)
   , (.=)
@@ -107,7 +113,7 @@ instance FromJSON Settings where
 defaultSettings :: Settings
 defaultSettings =
   Settings
-    { settingsVersion = 3
+    { settingsVersion = 4
     , audioDeviceName = Nothing
     , voiceThresholdDb = -42
     , trailingSilenceMs = 700
@@ -128,8 +134,142 @@ data HotkeyPreset
   | SuperShiftSpace
   | FunctionKey8
   | FunctionKey9
+  | CustomHotkey HotkeyModifiers HotkeyKey
+  deriving stock (Eq, Ord, Show, Generic)
+
+instance ToJSON HotkeyPreset where
+  toJSON preset = case preset of
+    ControlShiftSpace -> String "ControlShiftSpace"
+    ControlAltSpace -> String "ControlAltSpace"
+    SuperShiftSpace -> String "SuperShiftSpace"
+    FunctionKey8 -> String "FunctionKey8"
+    FunctionKey9 -> String "FunctionKey9"
+    CustomHotkey modifiers key ->
+      object
+        [ "tag" .= ("CustomHotkey" :: Text)
+        , "modifiers" .= modifiers
+        , "key" .= key
+        ]
+
+instance FromJSON HotkeyPreset where
+  parseJSON (String name) = case name of
+    "ControlShiftSpace" -> pure ControlShiftSpace
+    "ControlAltSpace" -> pure ControlAltSpace
+    "SuperShiftSpace" -> pure SuperShiftSpace
+    "FunctionKey8" -> pure FunctionKey8
+    "FunctionKey9" -> pure FunctionKey9
+    _ -> fail $ "Unknown shortcut preset: " <> Text.unpack name
+  parseJSON value = withObject "HotkeyPreset" parseCustom value
+    where
+      parseCustom o = do
+        tag <- o .: "tag"
+        if (tag :: Text) == "CustomHotkey"
+          then CustomHotkey <$> o .: "modifiers" <*> o .: "key"
+          else fail $ "Unknown shortcut type: " <> Text.unpack tag
+
+data HotkeyModifiers = HotkeyModifiers
+  { modifierControl :: Bool
+  , modifierShift :: Bool
+  , modifierAlt :: Bool
+  , modifierSuper :: Bool
+  }
   deriving stock (Eq, Ord, Show, Generic)
   deriving anyclass (FromJSON, ToJSON)
+
+data HotkeyKey
+  = HotkeyA
+  | HotkeyB
+  | HotkeyC
+  | HotkeyD
+  | HotkeyE
+  | HotkeyF
+  | HotkeyG
+  | HotkeyH
+  | HotkeyI
+  | HotkeyJ
+  | HotkeyK
+  | HotkeyL
+  | HotkeyM
+  | HotkeyN
+  | HotkeyO
+  | HotkeyP
+  | HotkeyQ
+  | HotkeyR
+  | HotkeyS
+  | HotkeyT
+  | HotkeyU
+  | HotkeyV
+  | HotkeyW
+  | HotkeyX
+  | HotkeyY
+  | HotkeyZ
+  | Hotkey0
+  | Hotkey1
+  | Hotkey2
+  | Hotkey3
+  | Hotkey4
+  | Hotkey5
+  | Hotkey6
+  | Hotkey7
+  | Hotkey8
+  | Hotkey9
+  | HotkeySpace
+  | HotkeyTab
+  | HotkeyReturn
+  | HotkeyEscape
+  | HotkeyBackspace
+  | HotkeyLeft
+  | HotkeyRight
+  | HotkeyUp
+  | HotkeyDown
+  | HotkeyHome
+  | HotkeyEnd
+  | HotkeyPageUp
+  | HotkeyPageDown
+  | HotkeyInsert
+  | HotkeyDelete
+  | HotkeyMinus
+  | HotkeyEquals
+  | HotkeyLeftBracket
+  | HotkeyRightBracket
+  | HotkeyBackslash
+  | HotkeySemicolon
+  | HotkeyQuote
+  | HotkeyBackquote
+  | HotkeyComma
+  | HotkeyPeriod
+  | HotkeySlash
+  | HotkeyF1
+  | HotkeyF2
+  | HotkeyF3
+  | HotkeyF4
+  | HotkeyF5
+  | HotkeyF6
+  | HotkeyF7
+  | HotkeyF8
+  | HotkeyF9
+  | HotkeyF10
+  | HotkeyF11
+  | HotkeyF12
+  deriving stock (Bounded, Enum, Eq, Ord, Show, Generic)
+  deriving anyclass (FromJSON, ToJSON)
+
+hotkeyBinding :: HotkeyPreset -> (HotkeyModifiers, HotkeyKey)
+hotkeyBinding preset = case preset of
+  ControlShiftSpace -> (HotkeyModifiers True True False False, HotkeySpace)
+  ControlAltSpace -> (HotkeyModifiers True False True False, HotkeySpace)
+  SuperShiftSpace -> (HotkeyModifiers False True False True, HotkeySpace)
+  FunctionKey8 -> (HotkeyModifiers False False False False, HotkeyF8)
+  FunctionKey9 -> (HotkeyModifiers False False False False, HotkeyF9)
+  CustomHotkey modifiers key -> (modifiers, key)
+
+validHotkeyBinding :: HotkeyModifiers -> HotkeyKey -> Bool
+validHotkeyBinding modifiers key =
+  modifierControl modifiers
+    || modifierShift modifiers
+    || modifierAlt modifiers
+    || modifierSuper modifiers
+    || key >= HotkeyF1
 
 data ActivationSource
   = HotkeyActivation